unify hash checking for rawmd4/rawmd5/nt-opencl.
